Add a check to see whether cmkr was run before pushing

+++ src/build.cpp
@@ -16,7 +16,7 @@ namespace build {
 int run(int argc, char **argv) {
     cmake::CMake cmake(".", true);
     if (argc > 2) {
-        for (size_t i = 2; i < argc; ++i) {
+        for (int i = 2; i < argc; ++i) {
             cmake.build_args.push_back(argv[i]);
         }
     }
